サインアウト URL の authUrl をサーバーで解決し client へ prop で渡す（相対 URL 化による signout 失敗を修正）
page.tsx をサーバーコンポーネント化し、NEXT_PUBLIC_AUTH_URL をランタイム env から
解決して HomePageClient に authUrl prop として渡す方式に変更する。
client component 内で process.env.NEXT_PUBLIC_AUTH_URL を参照すると Docker ビルド時に
空文字がインライン化され、buildSignOutUrl が相対 URL を返してサインアウトが失敗する問題を修正。

変更ファイル:
- services/livetalk/web/src/app/page.tsx: サーバーコンポーネントとして書き直し
- services/livetalk/web/src/app/HomePageClient.tsx: 旧 page.tsx の内容を移動し authUrl prop を追加
- services/livetalk/web/tests/unit/app/HomePageClient.test.tsx: authUrl prop の伝達を検証するテストを追加
